+++ klum-ast/src/test/groovy/com/blackbuild/groovy/configdsl/transform/SharedCapabilitiesDocumentaryTest.groovy
@@ -117,7 +117,7 @@ class SharedCapabilitiesDocumentaryTest extends AbstractDSLSpec {
                 @PostTree
                 void captureSpecialRegistryUrl() {
                     if (SpecialRegistry.Create.isBuilder(registry)) {
-                        def special = SpecialRegistry.Create.asBuilder(registry)
+                        def special = SpecialRegistry.Create.narrowBuilder(registry)
                         configuredRegistryUrl = special.toUrl()
                     }
                 }